Add JSON as a support for the room's objects
The objects aren't hard-coded into the room anymore. JSON files
with the list of objects are now supported as the standard format
to provide:

	-Object type
	-Coordinates

Therefore, I've implemented the scripts needed to fully decode the
JSON and a JSON file for the only level existent so far. Moreover,
the fuel and time bars are now created by the "main" object, obj_lvl.