Refactor drawable and saveable objs handling
Drawing instances to the screen in the liftoff stage no longer depend
on the instances' specific parentage: obj_liftoff_drawable. This
object served only as a grouping for the objects whose instances had
this 'draw' property.

Also, saveable instances that are part of the 'configuration' of a
level - the instances that define the level itself - didn't exist as
a separate grouping. The drawable object grouping served this
purpose instead.

Fix: #10